app_webscr_otodom/requirements.txt
pip
lxml
boto3
requests
pymongo
Pillow
pandas
botocore
numpy==1.15.4
xmltodict
Scrapy
twisted
pymongo